refactor!: move `Taxon` class into `common` package
This moves `Taxon` from being biodbnet-specific into a more package-wide `common` package

BREAKING-CHANGE: Import `Taxon` from `fast_bioservices.common` instead of `fast_bioservices.biodbnet`
