+++ src/fast_bioservices/biodbnet/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ["BioDBNet", "Input", "Output", "Taxon"]
+__all__ = ["BioDBNet", "Input", "Output"]
 
 from fast_bioservices.biodbnet.biodbnet import BioDBNet
-from fast_bioservices.biodbnet.nodes import Input, Output, Taxon
+from fast_bioservices.biodbnet.nodes import Input, Output
